requirements.txt
llama-index
llama-index-embeddings-jinaai
llama-index-llms-openai
Pillow
python-dotenv
fastapi[standard]
uvicorn
aiohttp
botbuilder-core
httpx